Use uint64_t for resume position
size_t is for data type size, not data size.

+++ lib/internal.h
@@ -119,8 +119,8 @@ typedef struct pndman_curl_handle
    void *curl;
    void *file;
    void *header_list;
-   size_t resume;
-   size_t retry;
+   uint64_t resume;
+   int retry;
    time_t if_modified_since;
    pndman_curl_callback callback;
    pndman_curl_progress *progress;
